requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==3.6.4.0
celery==5.2.7
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap5==2024.2
Deprecated==1.2.14
Django==5.1.1
django-crispy-forms==2.2
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
idna==3.7
kombu==5.3.7
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pytz==2024.1
redis==4.3.4
requests==2.32.3
sqlparse==0.5.0
stripe==9.12.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
